requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues


azure-functions
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
ffmpeg-python==0.2.0
future==0.18.3
idna==3.4
llvmlite==0.39.1
more-itertools==9.1.0
numba==0.56.4
numpy==1.23.5
openai-whisper @ git+https://github.com/openai/whisper.git@6dea21fd7f7253bfe450f1e2512a0fe47ee2d258
Pillow==9.4.0
regex==2022.10.31
requests==2.28.2
tiktoken==0.3.1
torch==1.8.1
torchvision==0.9.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
